feat: PostController REST API 구현
 커밋할 변경 사항:
	새 파일:       src/main/java/backend/airo/api/post/PostController.java
	새 파일:       src/main/java/backend/airo/api/post/dto/ErrorResponse.java
	새 파일:       src/main/java/backend/airo/api/post/dto/LikeResponse.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PopularPostsRequest.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostCreateRequest.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostDeleteRequest.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostDetailRequest.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostDetailResponse.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostListRequest.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostListResponse.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostResponse.java
	새 파일:       src/main/java/backend/airo/api/post/dto/PostUpdateRequest.java
	새 파일:       src/main/java/backend/airo/api/post/dto/RecentPopularPostsRequest.java
	수정함:        src/main/java/backend/airo/application/post/usecase/PostUseCase.java
	수정함:        src/main/java/backend/airo/domain/post/Post.java
	수정함:        src/main/java/backend/airo/domain/post/dto/PostSearchCriteria.java
	수정함:        src/main/java/backend/airo/domain/post/query/GetPostListQuery.java
	수정함:        src/main/java/backend/airo/domain/post/repository/PostRepository.java
	수정함:        src/main/java/backend/airo/persistence/post/adapter/PostAdapter.java
	수정함:        src/main/java/backend/airo/persistence/post/entity/PostEntity.java
	수정함:        src/main/java/backend/airo/persistence/post/repository/PostJpaRepository.java
	수정함:        src/main/java/backend/airo/security/config/SecurityConfig.java
	수정함:        src/main/resources/application.yml
	수정함:        src/test/java/backend/airo/application/post/usecase/PostUseCaseTest.java

+++ src/main/java/backend/airo/domain/post/dto/PostSearchCriteria.java
@@ -12,7 +12,7 @@ public record PostSearchCriteria(
         String searchScope, // title, content, all
         List<PostStatus> statuses,
         Long userId,
-        Integer categoryId,
+        Long categoryId,
         Long locationId,
         List<String> tags,
         Boolean isFeatured,
@@ -59,7 +59,7 @@ public record PostSearchCriteria(
      * 고급 검색 조건 생성
      */
     public static PostSearchCriteria advanced(String keyword, List<PostStatus> statuses,
-                                              Integer categoryId, List<String> tags) {
+                                              Long categoryId, List<String> tags) {
         return new PostSearchCriteria(
                 keyword, "all", statuses, null, categoryId, null,
                 tags, null, null, null, null, null, null, null, null
